report/pyproject.toml
[project]
name = "report"
version = "0.1.7"
description = "seeme-report"
authors = [
    {name = "lolik",email = "tuwork@foxmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pywin32 (>=308,<309)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ttkbootstrap (>=1.10.1,<2.0.0)"
]


[dependency-groups]
dev = [
    "nuitka>=2.8.4",
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 200
exclude = [".venv"]
